BUGFIX: introduce --enforce-target-formatting flag
This change fixes a bug where the current codegen will produce
duplicate keys in the output if they conflict during naming
formatting. This can lead to unexpected (and incorrect) schema
generation in the target language.

For example, currently, when using codegen from javascript to swift,
if you have an enum with keys `foo_bar` and `FOO_BAR`, the result
of the codegen with have an enum with two `fooBar` keys. This is
invalid swift, since swift enum cases cannot be redeclared.
This happens because codegen currently attempts to write output in the
style that is expected by developers of the target (which for swift
means turning keys into pascaleCase).

While producing the output in familiar styling is a great goal,
producing invalid output is not an acceptable default for most users.

This change changes the current default behavior to not attempt to
respect standard styling practices of the target. For user who still
want such styling, this change also adds a new flag,
`--enforce-target-formatting` which when `true` maintains the
current behavior.

The impetus for this change:
When Brigade first starded writing our schema, we wrote our enum
keys in pascalCase because we didn't know any better. However, we
learned quick enough that the recommended best practice for GQL was
to write enum keys in SCREAMING_CASE. So, we made a commit that
deprecated all camelCase enum keys and replaced them with
SCREAMING_CASE keys. Our node.js GraphQL server was perfectly happy
with this change. However, our mobile clients, which use codegen,
broke horribly. It took us hours to figure out why.

The result of the above investigation was this conversation a bit
over half a year ago:
https://apollographql.slack.com/archives/C6FNM649M/p1509129702000403

This commit does not follow the instructions of that conversation.
After attending Apollo Day last week and talking with many Apollo
employees/devs, it was apparent to me that in general, Apollo
products should "just work." This commit is in that spirit.

It should be further noted that the developer running codegen may
very well have no control over the schema. Thus it is imperative
that the output of codegen always be correct in the default case.