BREAKING: introduce --enforce-target-formatting flag
This change fixes a bug where the current codegen will drop keys
from the schema if they conflict during naming formatting. This
leads to unexpected (and incorrect) schema generation in the target
language.

For example, currently, when using codegen from javascript to swift,
if you have an enum with keys `foo_bar` and `FOO_BAR`, the result
of the codegen with have an enum with only the single key, `fooBar`.
This is because codegen currently attempts to write output in the
style that is expected by developers of the target.

While the above is a great goal, producing invalid output (relative
to the original schema) is not an acceptable default for most users.

This change changes the current default behavior to not attempt to
respect standard styling practices of the target. For user who still
want such styling, or users who want to upgrade without having to
change lots of code, this change also adds a new flag,
`--enforce-target-formatting` which when `true` maintains the
current behavior of potentially overriding keys.

The impetus for this change:
When Brigade first starded writing our schema, we wrote our enum
keys in pascalCase because we didn't know any better. However, we
learned quick enough the recommended best practice for GQL was to
write enum keys in SCREAMING_CASE. So, we made a commot that
deprecated all camelCase enum keys and replaced them with
SCREAMING_CASE keys. Our node.js GraphQL server was perfectly happy
with this change. However, our mobile clients, which use codegen,
broke horribly. It took us hours to figure out why.

The result of the above investigation was this conversation a bit
over half a year ago:
https://apollographql.slack.com/archives/C6FNM649M/p1509129702000403

This commit does not follow the instructions of that conversation.
After attending Apollo Day last week and talking with many Apollo
employees/devs, it was apparent to me that in general, Apollo
products should "just work." This commit is in that spirit.